Move master list input filter and expose query by master list
