+++ server/graphql/types/src/types/master_list.rs
@@ -1,11 +1,46 @@
-use async_graphql::{Context, Error, Object};
-use graphql_core::{standard_graphql_error::StandardGraphqlError, ContextExt};
-use repository::MasterList;
+use async_graphql::{Context, Error, InputObject, Object};
+use graphql_core::{
+    generic_filters::{EqualFilterStringInput, StringFilterInput},
+    standard_graphql_error::StandardGraphqlError,
+    ContextExt,
+};
+use repository::{EqualFilter, MasterList, MasterListFilter, StringFilter};
 use service::master_list::query_lines::get_master_list_lines_count;
 
 #[derive(PartialEq, Debug)]
 pub struct MasterListNode {
     master_list: MasterList,
+}
+
+#[derive(InputObject, Clone)]
+pub struct MasterListFilterInput {
+    pub id: Option<EqualFilterStringInput>,
+    pub name: Option<StringFilterInput>,
+    pub code: Option<StringFilterInput>,
+    pub description: Option<StringFilterInput>,
+    pub exists_for_name: Option<StringFilterInput>,
+    pub exists_for_name_id: Option<EqualFilterStringInput>,
+    pub exists_for_store_id: Option<EqualFilterStringInput>,
+    pub is_program: Option<bool>,
+    pub item_id: Option<EqualFilterStringInput>,
+}
+
+impl MasterListFilterInput {
+    pub fn to_domain(self) -> MasterListFilter {
+        MasterListFilter {
+            id: self.id.map(EqualFilter::from),
+            name: self.name.map(StringFilter::from),
+            code: self.code.map(StringFilter::from),
+            description: self.description.map(StringFilter::from),
+            exists_for_name: self.exists_for_name.map(StringFilter::from),
+            exists_for_name_id: self.exists_for_name_id.map(EqualFilter::from),
+            exists_for_store_id: self.exists_for_store_id.map(EqualFilter::from),
+            is_program: self.is_program,
+            item_id: self.item_id.map(EqualFilter::from),
+            is_discount_list: None,
+            is_default_price_list: None,
+        }
+    }
 }
 
 #[Object]
